--- t/smoke/common/__init__.py
@@ -1,2 +1,0 @@
-from t.smoke.common.celery4 import *  # noqa
-from t.smoke.common.worker import *  # noqa
