--- t/smoke/common/celery4/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-
-
-from t.smoke.common.celery4.api import Celery4WorkerContainer
-from t.smoke.common.celery4.fixtures import *  # noqa
