--- t/smoke/common/worker/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-
-
-from t.smoke.common.worker.api import SmokeWorkerContainer
-from t.smoke.common.worker.fixtures import *  # noqa
