+++ t/smoke/conftest.py
@@ -4,14 +4,14 @@ import pytest
 from pytest_celery.api.components.worker.cluster import CeleryWorkerCluster
 from pytest_celery.api.components.worker.node import CeleryTestWorker
 
-from t.smoke.common import *  # noqa
+from t.smoke.workers import *  # noqa
 
 
 @pytest.fixture(
     # Each param item is a list of workers to be used in the cluster
     params=[
         ["celery_setup_worker"],
-        # ["celery_setup_worker", "celery4_worker"],
+        ["celery_setup_worker", "celery4_worker"],
     ]
 )
 def celery_worker_cluster(request: pytest.FixtureRequest) -> CeleryWorkerCluster:
