+++ t/smoke/test_canvas.py
@@ -4,7 +4,7 @@ from pytest_celery.api.components.worker.node import CeleryTestWorker
 from pytest_docker_tools.wrappers.container import wait_for_callable
 
 from celery.canvas import chain, chord, group, signature
-from t.smoke.common.tasks import add, identity
+from t.smoke.tasks import add, identity
 
 
 class test_canvas:
