+++ t/smoke/workers/__init__.py
@@ -1,0 +1,2 @@
+from t.smoke.workers.celery4 import *  # noqa
+from t.smoke.workers.dev import *  # noqa
