+++ t/smoke/workers/celery4/Dockerfile
