+++ t/smoke/workers/celery4/fixtures.py
@@ -5,25 +5,13 @@ from pytest_celery.containers.worker import CeleryWorkerContainer
 from pytest_docker_tools import build, container, fxtr
 
 from celery import Celery
-from t.smoke.common.celery4.api import Celery4WorkerContainer
+from t.smoke.workers.celery4.api import Celery4WorkerContainer
 
 celery4_worker_image = build(
-    path="t/smoke/common/celery4",
+    path="t/smoke/workers/celery4",
     tag="t/smoke/worker:celery4",
     buildargs=Celery4WorkerContainer.buildargs(),
 )
-
-
-@pytest.fixture
-def celery4_worker(
-    celery4_worker_container: CeleryWorkerContainer,
-    celery_setup_app: Celery,
-) -> CeleryTestWorker:
-    worker = CeleryTestWorker(
-        celery4_worker_container,
-        app=celery_setup_app,
-    )
-    yield worker
 
 
 celery4_worker_container = container(
@@ -34,3 +22,8 @@ celery4_worker_container = container(
     wrapper_class=Celery4WorkerContainer,
     timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
 )
+
+
+@pytest.fixture
+def celery4_worker(celery4_worker_container: CeleryWorkerContainer, celery_setup_app: Celery) -> CeleryTestWorker:
+    yield CeleryTestWorker(celery4_worker_container, app=celery_setup_app)
