+++ t/smoke/workers/dev/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+
+
+from t.smoke.workers.dev.api import SmokeWorkerContainer
+from t.smoke.workers.dev.fixtures import *  # noqa
