+++ t/smoke/workers/dev/api.py
@@ -2,15 +2,17 @@ from typing import Any
 
 from pytest_celery.containers.worker import CeleryWorkerContainer
 
+import celery
 
-class Celery4WorkerContainer(CeleryWorkerContainer):
+
+class SmokeWorkerContainer(CeleryWorkerContainer):
     @property
     def client(self) -> Any:
         return self
 
     @classmethod
     def version(cls) -> str:
-        return "4.4.7"
+        return celery.__version__
 
     @classmethod
     def log_level(cls) -> str:
@@ -18,8 +20,8 @@ class Celery4WorkerContainer(CeleryWorkerContainer):
 
     @classmethod
     def worker_name(cls) -> str:
-        return "celery4_worker"
+        return "smoke_tests_worker"
 
     @classmethod
     def worker_queue(cls) -> str:
-        return "celery4"
+        return "smoke_tests_queue"
