+++ t/smoke/workers/dev/fixtures.py
@@ -6,30 +6,29 @@ from pytest_celery import defaults
 from pytest_celery.containers.worker import CeleryWorkerContainer
 from pytest_docker_tools import build, container, fxtr
 
-from t.smoke.common.worker.api import SmokeWorkerContainer
+from t.smoke.workers.dev.api import SmokeWorkerContainer
 
 smoke_worker_image = build(
     path=".",
-    dockerfile="t/smoke/common/worker/Dockerfile",
+    dockerfile="t/smoke/workers/dev/Dockerfile",
     tag="t/smoke/worker:dev",
     buildargs=SmokeWorkerContainer.buildargs(),
 )
 
-worker_volume = {
-    # Volume: Worker /app
-    "{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME,
-    # Mount: Celery source
-    os.path.abspath(os.getcwd()): {
-        "bind": "/celery",
-        "mode": "rw",
-    },
-}
 
 default_worker_container = container(
     image="{smoke_worker_image.id}",
     environment=fxtr("default_worker_env"),
     network="{DEFAULT_NETWORK.name}",
-    volumes=worker_volume,
+    volumes={
+        # Volume: Worker /app
+        "{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME,
+        # Mount: Celery source
+        os.path.abspath(os.getcwd()): {
+            "bind": "/celery",
+            "mode": "rw",
+        },
+    },
     wrapper_class=SmokeWorkerContainer,
     timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
 )
@@ -47,8 +46,8 @@ def default_worker_container_session_cls() -> Type[CeleryWorkerContainer]:
 
 @pytest.fixture
 def default_worker_tasks() -> set:
-    from t.smoke.common import tasks as common_tasks
+    from t.smoke import tasks as smoke_tests_tasks
 
     yield {
-        common_tasks,
+        smoke_tests_tasks,
     }
